+++ crates/lib/core/src/handlers/mod.rs
@@ -4,6 +4,7 @@ use core::ops::Range;
 use miden_core::Felt;
 use miden_processor::ProcessorState;
 
+pub mod aead_eidos;
 pub mod debug;
 pub mod ecdsa_k256_keccak;
 pub mod falcon_div;